+++ packages/cli/src/commands/firewall/ip-blocks/unblock.ts
@@ -140,6 +140,7 @@ export default async function unblock(client: Client, argv: string[]) {
       {
         action: 'ip.remove',
         id: rule.id,
+        value: null,
       },
       { teamId }
     );